+++ src/include/nodes/makefuncs.h
@@ -27,7 +27,7 @@ extern Var *
 makeVar(Index varno,
 		AttrNumber varattno,
 		Oid vartype,
-		int16 vartypmod,
+		int32 vartypmod,
 		Index varlevelsup,
 		Index varnoold,
 		AttrNumber varoattno);
@@ -35,7 +35,7 @@ makeVar(Index varno,
 extern Resdom *
 makeResdom(AttrNumber resno,
 		   Oid restype,
-		   int16 restypmod,
+		   int32 restypmod,
 		   char *resname,
 		   Index reskey,
 		   Oid reskeyop,
